Added Map to UserCommand context.
Map and Mappable interactions make a little more sense. If a Mappable
moves to a location, its map is set to the location's map if it is not the
same map as the Mappable's. Being directly removed from a Map via
Map.removeFromContents will also move the Mappable to empty space (null
location and map).